requirements.txt
fastapi
uvicorn
pymongo
PyJWT
python-dotenv